backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
cryptography==41.0.4
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
dict2xml==1.7.3
dill==0.3.7
dump==0.0.5
ecdsa==0.18.0
executing==2.0.0
f5-icontrol-rest==1.3.13
fastapi==0.103.2
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
genie==23.9
genie.libs.clean==23.9
genie.libs.conf==23.9
genie.libs.filetransferutils==23.9
genie.libs.health==23.9
genie.libs.ops==23.9
genie.libs.parser==23.9
genie.libs.sdk==23.9
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jupyter_client==8.3.1
jupyter_core==5.3.2
langchain==0.0.307
langsmith==0.0.41
lxml==4.9.3
mangum==0.17.0
marshmallow==3.20.1
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
ncclient==0.6.15
nest-asyncio==1.5.8
netaddr==0.9.0
numpy==1.26.0
onnxruntime==1.16.0
openai==0.28.1
overrides==7.4.0
packaging==23.2
paramiko==3.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
ply==3.11
posthog==3.0.2
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyftpdlib==1.5.8
Pygments==2.16.1
PyNaCl==1.5.0
pypdf==3.16.4
PyPika==0.48.9
pysmi-lextudio==1.1.13
pysnmp-lextudio==5.0.29
pysnmpcrypto==0.0.4
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
rest.connector==23.9
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tftpy==0.8.0
tiktoken==0.5.1
tokenizers==0.14.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
unicon==23.9
unicon.plugins==23.9
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
xmltodict==0.13.0
yang.connector==23.9
yarl==1.9.2
#End Requirements